seperated out the project types definition file from the hardhat.config.ts and swapped out relative paths to an alias.
