+++ store/contracts.ts
@@ -1,5 +1,5 @@
 import { Module, VuexModule } from "vuex-module-decorators"
-import { LSPConfiguration } from "~/hardhat.config"
+import { LSPConfiguration } from "~/types"
 
 @Module({
   stateFactory: true,
@@ -7,7 +7,7 @@ import { LSPConfiguration } from "~/hardhat.config"
   namespaced: true,
 })
 export default class contracts extends VuexModule {
-  contractConfigs: Array<LSPConfiguration> = require("../deployedContractConfigs.json")
+  contractConfigs: Array<LSPConfiguration> = require("~/deployedContractConfigs.json")
 
   get syntheticNames() {
     return this.contractConfigs.map((config) => config.syntheticName)
